Refactor script normalization to improve maintainability
